feat: make default setup more multibrand friendly
in order to enable multibrand, the files in src/templates
can be copied directly into subfolders for specific domains
previously this would have caused lots of duplication and
need awkward repetitive refactoring. now the files should
only need minor edits to specify any custom filters.

J=CR-2420

TEST=none
